pyproject.toml
[project]
name = "describe-act"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4.8",
    "pyyaml>=6.0.2",
]
